pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum_cognition"
version = "0.1.0"
description = "Kurzbeschreibung des Projekts"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Dein Name"}]
dependencies = [
  "python-dotenv>=1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "black>=24.0",
  "isort>=5.13",
  "pre-commit>=3.7",
  "ipykernel>=6.29",
  "pyarrow>=17.0"
]

[tool.black]
line-length = 100
target-version = ["py311"]
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-select = ["I"]  # Import-Ordnung via Ruff
src = ["src", "tests"]
exclude = [".venv"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
